First pass removal of TypeResolver
